// frontend/src/shared/api/model/adminUserResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface AdminUserResponse {
  id?: number;
  fullName?: string;
  email?: string;
  phone?: string;
  verified?: boolean;
  enabled?: boolean;
  createdAt?: string;
  roles?: string[];
}